Log request-id if present
tracing from service to is difficult so adding request-id helps in
cross-referencing in logs.

Furthermore, we get when the request is received and then our nginx
logs log when the request is handled so we can trace potential
bottlenecks in service  -> paper-pusher -> service -> paper-pusher ->
service shenanigans of getting the skeleton for a fillable pdf.

for those playing along at home:
    service (ask for generator info)
 -> paper-pusher (pdf url)
 -> service (s3 pdf location)
 -> paper-pusher (pdf bytes and parses to get skeleton)
 -> service

Co-authored-by: Sean Vincent <sean.vincent@aclaimant.com>